Remove uniqkey field from every model.
This field is redundant data. Every Django database model has a built-in
id which is unique for that database entity. So this field becomes
redundant. In addition this field is not even a primary key. So I am
not sure what it is trying to achieve. It is also agains the schema
design principles. Normalization level 2 is lost by having redundant
data.

The way uniqkey is generated using random is also awkward.

--HG--
extra : rebase_source : f02c906ecc68657d554cf3fb2da471252a4f62dd