Create a package for taskapp views and break the views into task and textbook.
Now all the view functions common to any two entities along with all
tasks related views sit in task module. Even if the view is not directly
related to the task entity, it sits in the task module since task is the
base for every other entity in the application.